Factor out template input handling into common package
This code will be needed for the `upgrade` command, so it can't live with
the `render` command anymore.

This involved:
 - Turning methods into plain methods. This required passing a lot more
   parameters around, so we create the `ResolveParams` struct to
   encapsulate them.
 - Moving code from render.go into a new package, common/input/input.go.
   Along with associated tests.
