+++ templates/commands/render/render_test.go
@@ -16,7 +16,6 @@ package render
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"io/fs"
@@ -29,9 +28,9 @@ import (
 	"github.com/benbjohnson/clock"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"golang.org/x/exp/maps"
 
 	"github.com/abcxyz/abc/templates/common"
+	"github.com/abcxyz/abc/templates/common/input"
 	"github.com/abcxyz/abc/templates/model"
 	spec "github.com/abcxyz/abc/templates/model/spec/v1beta2"
 	"github.com/abcxyz/pkg/cli"
@@ -949,7 +948,7 @@ func testMustGlob(t *testing.T, glob string) (string, bool) {
 	panic("unreachable") // silence compiler warning for "missing return"
 }
 
-func TestPromptForInputs(t *testing.T) {
+func TestPromptDialog(t *testing.T) {
 	t.Parallel()
 
 	type dialogStep struct {
@@ -1216,11 +1215,6 @@ Enter value, or leave empty to accept default: `,
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
-			inputs := map[string]string{}
-			if tc.flagInputVals != nil {
-				inputs = maps.Clone(tc.flagInputVals)
-			}
-
 			cmd := &Command{}
 
 			stdinReader, stdinWriter := io.Pipe()
@@ -1233,11 +1227,21 @@ Enter value, or leave empty to accept default: `,
 
 			ctx := context.Background()
 			errCh := make(chan error)
+			var got map[string]string
 			go func() {
 				defer close(errCh)
-				errCh <- cmd.promptForInputs(ctx, &spec.Spec{
-					Inputs: tc.inputs,
-				}, inputs)
+				params := &input.ResolveParams{
+					Inputs:             tc.flagInputVals,
+					Prompt:             true,
+					Prompter:           cmd,
+					SkipPromptTTYCheck: true,
+					Spec:               &spec.Spec{Inputs: tc.inputs},
+				}
+				var err error
+				fmt.Printf("to input.Resolve\n")
+				got, err = input.Resolve(ctx, params)
+				fmt.Printf("from input.Resolve\n")
+				errCh <- err
 			}()
 
 			for _, ds := range tc.dialog {
@@ -1253,330 +1257,8 @@ Enter value, or leave empty to accept default: `,
 			case <-time.After(time.Second):
 				t.Fatal("timed out waiting for background goroutine to finish")
 			}
-			if diff := cmp.Diff(inputs, tc.want, cmpopts.EquateEmpty()); diff != "" {
+			if diff := cmp.Diff(got, tc.want, cmpopts.EquateEmpty()); diff != "" {
 				t.Fatalf("input values were different than expected (-got,+want): %s", diff)
-			}
-		})
-	}
-}
-
-func TestPromptForInputs_CanceledContext(t *testing.T) {
-	t.Parallel()
-
-	cmd := &Command{
-		flags: RenderFlags{
-			Inputs: map[string]string{},
-		},
-	}
-
-	stdinReader, _ := io.Pipe()
-	stdoutReader, stdoutWriter := io.Pipe()
-	_, stderrWriter := io.Pipe()
-
-	cmd.SetStdin(stdinReader)
-	cmd.SetStdout(stdoutWriter)
-	cmd.SetStderr(stderrWriter)
-
-	ctx, cancel := context.WithCancel(context.Background())
-
-	errCh := make(chan error)
-	go func() {
-		defer close(errCh)
-		spec := &spec.Spec{
-			Inputs: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-				},
-			},
-		}
-		errCh <- cmd.promptForInputs(ctx, spec, map[string]string{})
-	}()
-
-	go func() {
-		for {
-			// Read and discard prompt printed to the user.
-			if _, err := stdoutReader.Read(make([]byte, 1024)); err != nil {
-				return
-			}
-		}
-	}()
-
-	cancel()
-	var err error
-	select {
-	case err = <-errCh:
-	case <-time.After(time.Second):
-		t.Fatal("timed out waiting for the background goroutine to finish")
-	}
-
-	if !errors.Is(err, context.Canceled) {
-		t.Fatalf("got an error %v, want context.Canceled", err)
-	}
-
-	stdoutWriter.Close() // terminate the background goroutine blocking on stdoutReader.Read()
-}
-
-func TestValidateInputs(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		name        string
-		inputModels []*spec.Input
-		inputVals   map[string]string
-		want        string
-	}{
-		{
-			name: "no-validation-rule",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-		},
-		{
-			name: "single-passing-validation-rule",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule:    model.String{Val: `size(my_input) < 5`},
-							Message: model.String{Val: "Length must be less than 5"},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-		},
-		{
-			name: "single-failing-validation-rule",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule:    model.String{Val: `size(my_input) < 3`},
-							Message: model.String{Val: "Length must be less than 3"},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         size(my_input) < 3
-Rule msg:     Length must be less than 3`,
-		},
-		{
-			name: "multiple-passing-validation-rules",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule:    model.String{Val: `size(my_input) < 5`},
-							Message: model.String{Val: "Length must be less than 5"},
-						},
-						{
-							Rule:    model.String{Val: `my_input.startsWith("fo")`},
-							Message: model.String{Val: `Must start with "fo"`},
-						},
-						{
-							Rule:    model.String{Val: `my_input.contains("oo")`},
-							Message: model.String{Val: `Must contain "oo"`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-		},
-		{
-			name: "multiple-passing-validation-rules-one-failing",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule:    model.String{Val: `size(my_input) < 3`},
-							Message: model.String{Val: "Length must be less than 3"},
-						},
-						{
-							Rule:    model.String{Val: `my_input.startsWith("fo")`},
-							Message: model.String{Val: `Must start with "fo"`},
-						},
-						{
-							Rule:    model.String{Val: `my_input.contains("oo")`},
-							Message: model.String{Val: `Must contain "oo"`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         size(my_input) < 3
-Rule msg:     Length must be less than 3`,
-		},
-		{
-			name: "multiple-failing-validation-rules",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule:    model.String{Val: `size(my_input) < 3`},
-							Message: model.String{Val: "Length must be less than 3"},
-						},
-						{
-							Rule:    model.String{Val: `my_input.startsWith("ham")`},
-							Message: model.String{Val: `Must start with "ham"`},
-						},
-						{
-							Rule:    model.String{Val: `my_input.contains("shoe")`},
-							Message: model.String{Val: `Must contain "shoe"`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         size(my_input) < 3
-Rule msg:     Length must be less than 3
-
-Input name:   my_input
-Input value:  foo
-Rule:         my_input.startsWith("ham")
-Rule msg:     Must start with "ham"
-
-Input name:   my_input
-Input value:  foo
-Rule:         my_input.contains("shoe")
-Rule msg:     Must contain "shoe"`,
-		},
-		{
-			name: "cel-syntax-error",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule: model.String{Val: `(`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         (
-CEL error:    failed compiling CEL expression: ERROR: <input>:1:2: Syntax error:`, // remainder of error omitted
-		},
-		{
-			name: "cel-type-conversion-error",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule: model.String{Val: `bool(42)`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         bool(42)
-CEL error:    failed compiling CEL expression: ERROR: <input>:1:5: found no matching overload for 'bool'`, // remainder of error omitted
-		},
-		{
-			name: "cel-output-type-conversion-error",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule: model.String{Val: `42`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input": "foo",
-			},
-			want: `input validation failed:
-
-Input name:   my_input
-Input value:  foo
-Rule:         42
-CEL error:    CEL expression result couldn't be converted to bool. The CEL engine error was: unsupported type conversion from 'int' to bool`, // remainder of error omitted
-		},
-		{
-			name: "multi-input-validation",
-			inputModels: []*spec.Input{
-				{
-					Name: model.String{Val: "my_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule: model.String{Val: `my_input + my_other_input == "sharknado"`},
-						},
-					},
-				},
-				{
-					Name: model.String{Val: "my_other_input"},
-					Rules: []*spec.InputRule{
-						{
-							Rule: model.String{Val: `"tor" + my_other_input + my_input == "tornadoshark"`},
-						},
-					},
-				},
-			},
-			inputVals: map[string]string{
-				"my_input":       "shark",
-				"my_other_input": "nado",
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			r := &Command{}
-			ctx := context.Background()
-			err := r.validateInputs(ctx, tc.inputModels, tc.inputVals)
-			if diff := testutil.DiffErrString(err, tc.want); diff != "" {
-				t.Error(diff)
 			}
 		})
 	}
